fix: 修正 class插件 加载不完全
