+++ package.json
@@ -30,6 +30,7 @@
     "typescript": "^5.7.3"
   },
   "peerDependencies": {
+    "node-karin": "workspace:*",
     "@karinjs/karin-types": "workspace:*"
   }
 }
