+++ packages/core/src/plugin/index.ts
@@ -1,7 +1,5 @@
 import { LoaderPlugin } from './loader'
 
-export * from './list'
-
 /**
  * @internal
  * @description 初始化插件
